+++ utils/utils-internal/src/main/scala/pl/touk/nussknacker/engine/util/loader/LoadClassFromClassLoader.scala
@@ -17,6 +17,8 @@ trait LoadClassFromClassLoader {
         throw new IllegalArgumentException(s"$prettyClassName not found. ${jarsUrlsHint(classLoader)}")
       case Many(muchEntities) =>
         throw new IllegalArgumentException(s"Many $prettyClassName implementations found. Classes found: $muchEntities. ${jarsUrlsHint(classLoader)}")
+      //        I've found it the easiest way to run tests in IntelliJ when there's wrong classpath
+      //        muchEntities.last
       case One(only) => only
     }
   }